requirements/dev.txt
python-dotenv==0.19.2
black==22.1.0
pre-commit==2.9.3
pytest==7.1.1
pytest-mock==3.5.1
requests-mock==1.8.0
pytest-bdd==5.0.0